convert the rest of usize IDs to u32. savings:
maps: 645mb -> 625mb
scenarios: 431mb -> 390mb (before all the u32 optimizations, this as
500mb!!!)
prebaked: 80mb -> 73mb

and while I'm at it, grab fresh OSM, with lots of manual lane fixes,
especially near divided highways

+++ game/src/common/warp.rs
@@ -104,9 +104,9 @@ fn inner_warp(ctx: &mut EventCtx, app: &mut App, line: &str) -> Option<Transitio
                 let r = app.primary.map.maybe_get_r(RoadID(idx))?;
                 ID::Lane(r.children_forwards[0].0)
             }
-            'l' => ID::Lane(LaneID(idx as u32)),
+            'l' => ID::Lane(LaneID(idx)),
             'i' => ID::Intersection(IntersectionID(idx)),
-            'b' => ID::Building(BuildingID(idx as u32)),
+            'b' => ID::Building(BuildingID(idx)),
             'a' => ID::Area(AreaID(idx)),
             'p' => ID::Pedestrian(PedestrianID(idx)),
             'P' => {
